Caching Infer.

+++ .resource/script/infer.sh
@@ -11,14 +11,16 @@ ls -la ${HOME}
 ls -la $INFER_HOME
 mkdir -p $INFER_HOME
 
-# Download last Infer:
-curl -sSL $INFER_URL | tar -C $INFER_HOME -xJ
+# Install Infer if not exists:
+if [ ! -f "$INFER_HOME/bin/infer" ]; then
+	# Download last Infer:
+	curl -sSL $INFER_URL | tar -C $INFER_HOME -xJ
 
-# Move one level up:
-mv $INFER_HOME/infer-linux64-v$INFER_VERSION/* .
-rm -fr $INFER_HOME/infer-linux64-v$INFER_VERSION
+	# Move one level up:
+	mv $INFER_HOME/infer-linux64-v$INFER_VERSION/* $INFER_HOME
+	rm -fr $INFER_HOME/infer-linux64-v$INFER_VERSION
+fi
 
-ls -la ${HOME}
 ls -la $INFER_HOME
 
 # Install Python v2.7 (required by Infer):
